backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-tenants==3.8.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
cryptography==41.0.8
